# lib/api-zod/src/generated/types/annualFinancial.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Market Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface AnnualFinancial {
  year: string;
  revenue?: number | null;
  netIncome?: number | null;
  eps?: number | null;
  operatingIncome?: number | null;
  grossProfit?: number | null;
}